My Datapacks
Veinmeiner: Recreates an exclusively modded version of an additional gameplay feature. Allows the breaking of every type of a block connected to the broken to be broken all at once.
Magic Datapck: Adds several magical staffs and new recipies along with some secrets.
Leveling Datapack: Adds leveling systems to minecraft similar to an rpg that increase loot based off of your proficiency in the task required to get that loot.